move type benchmark to scylla-cql
So that it can access its full types API.
